# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Homing API",
    "version": "1.0.0-rc.1",
    "description": "Machine-readable route, method, authentication, and common-envelope contract. Field-level invariants consumed by the unchanged Python client remain locked by its real-client acceptance suite."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v1/me/token": {
      "get": {
        "operationId": "get_api_v1_me_token",
        "summary": "GET /api/v1/me/token",
        "tags": [
          "agent-required"
        ],
        "security": [
          {
            "cookieSession": []
          },
          {
            "bearerAuth": []
          }
        ],
        "x-homing-classification": "agent-required",
        "x-homing-csrf": false,
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonValue"
                }
              }
            }
          },
          "default": {
            "description": "Standard Homing error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete_api_v1_me_token",
        "summary": "DELETE /api/v1/me/token",
        "tags": [
          "agent-required"
        ],
        "security": [
          {
            "cookieSession": []
          },
          {
            "bearerAuth": []
          }
        ],
        "x-homing-classification": "agent-required",
        "x-homing-csrf": false,
        "responses": {
          "204": {
            "description": "No content",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "description": "Standard Homing error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/token/finalize-setup": {
      "post": {
        "operationId": "post_api_v1_me_token_finalize_setup",
        "summary": "POST /api/v1/me/token/finalize-setup",
        "tags": [
          "agent-required"
        ],
        "security": [
          {
            "cookieSession": []
          },
          {
            "bearerAuth": []
          }
        ],
        "x-homing-classification": "agent-required",
        "x-homing-csrf": false,
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonValue"
                }
              }
            }
          },
          "default": {
            "description": "Standard Homing error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/agent-link": {
      "post": {
        "operationId": "post_api_v1_agent_link",
        "summary": "POST /api/v1/agent-link",
        "tags": [
          "agent-required"
        ],
        "security": [],
        "x-homing-classification": "agent-required",
        "x-homing-csrf": false,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JsonObject"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonValue"
                }
              }
            }
          },
          "default": {
            "description": "Standard Homing error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent-link/token": {
      "post": {
        "operationId": "post_api_v1_agent_link_token",
        "summary": "POST /api/v1/agent-link/token",
        "tags": [
          "agent-required"
        ],
        "security": [],
        "x-homing-classification": "agent-required",
        "x-homing-csrf": false,
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JsonObject"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonValue"
                }
              }
            }
          },
          "default": {
            "description": "Standard Homing error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agent/projects": {
      "get": {
        "operationId": "get_api_v1_agent_projects",
        "summary": "GET /api/v1/agent/projects",
        "tags": [
          "agent-required"
        ],
        "security": [
          {
            "cookieSession": []
          },
          {
            "bearerAuth": []
          }
        ],
        "x-homing-classification": "agent-required",
        "x-homing-csrf": false,
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonValue"
                }
              }
            }
          },
          "default": {
            "description": "Standard Homing error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project}/config-revisions": {
      "post": {
        "operationId": "post_api_v1_projects_project_config_revisions",
        "summary": "POST /api/v1/projects/{project}/config-revisions",
        "tags": [
          "agent-required"
        ],
        "security": [
          {
            "cookieSession": []
          },
          {
            "bearerAuth": []
          }
        ],
        "x-homing-classification": "agent-required",
        "x-homing-csrf": false,
        "parameters": [
          {
            "in": "path",
            "name": "project",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JsonObject"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonValue"
                }
              }
            }
          },
          "default": {
            "description": "Standard Homing error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project}/config-revisions/{revision}": {
      "get": {
        "operationId": "get_api_v1_projects_project_config_revisions_revision",
        "summary": "GET /api/v1/projects/{project}/config-revisions/{revision}",
        "tags": [
          "agent-required"
        ],
        "security": [
          {
            "cookieSession": []
          },
          {
            "bearerAuth": []
          }
        ],
        "x-homing-classification": "agent-required",
        "x-homing-csrf": false,
        "parameters": [
          {
            "in": "path",
            "name": "project",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "path",
            "name": "revision",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonValue"
                }
              }
            }
          },
          "default": {
            "description": "Standard Homing error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project}/source-query-revisions/{query}": {
      "get": {
        "operationId": "get_api_v1_projects_project_source_query_revisions_query",
        "summary": "GET /api/v1/projects/{project}/source-query-revisions/{query}",
        "tags": [
          "agent-required"
        ],
        "security": [
          {
            "cookieSession": []
          },
          {
            "bearerAuth": []
          }
        ],
        "x-homing-classification": "agent-required",
        "x-homing-csrf": false,
        "parameters": [
          {
            "in": "path",
            "name": "project",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "path",
            "name": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "headers": {
              "X-Request-ID": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonValue"
                }
              }
            }
          },
          "default": {
            "description": "Standard Homing error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-homing_session"
      }
    },
    "schemas": {
      "JsonObject": {
        "type": "object",
        "additionalProperties": true
      },
      "JsonValue": {},
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "request_id"
            ],
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "fields": {
                "type": "object",
                "additionalProperties": true
              },
              "request_id": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  }
}